Change the data type of seconds in IQ:last from int to Long
Making it Long allows it to hold an XML-unsignedLong value as well
as null values. Before this patch, it was an int and defaulted to 0.
This was not right as int is too small to hold number of seconds for
last activity time and primitive data types do not allow for null values.

Test-information:

tested using an XMPP client to query last IQ on MUC rooms

Reviewer: Kevin Smith <kevin.smith@isode.com>

+++ src/com/isode/stroke/elements/Last.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, Isode Limited, London, England.
+ * Copyright (c) 2011-2013, Isode Limited, London, England.
  * All rights reserved.
  */
 /*
@@ -10,19 +10,19 @@
 package com.isode.stroke.elements;
 
 public class Last extends Payload {
-    int seconds_;
+    Long seconds_;
 
     public Last() {}
 
-    public Last(final int seconds) {
+    public Last(final Long seconds) {
         setSeconds(seconds);
     }
 
-    public void setSeconds(final int seconds) {
+    public void setSeconds(final Long seconds) {
         seconds_ = seconds;
     }
 
-    public int getSeconds() {
+    public Long getSeconds() {
         return seconds_;
     }
 }
